refactor(api): modularize category controller and remove unnecesary imports
