+++ api/src/controllers/index.js
@@ -1,8 +1,8 @@
 const getEvents = require('./event/getEvents')
 const { signupUser, loginUser, protectedUser, } = require('./user')
 const { createEvent } = require('./event/createEvent')
-const getCategories = require('./getCategories')
-const { getEventsByCategory } = require('./getEventsByCategory')
+const getCategories = require('./category/getCategories')
+const { getEventsByCategory } = require('./category/getEventsByCategory')
 const setCountries = require('./location/setCountries')
 const getCountries = require('./location/getCountries')
 const setStates = require('./location/setStates')
